+++ tests/run.clj
@@ -59,6 +59,7 @@
 (require "tests/clj_metadata_test")
 (require "tests/empty_list_test")
 (require "tests/error_path_test")
+(require "tests/bc_try_catch_test")
 (require "tests/regex_reentrant_test")
 (require "tests/ifn_test")
 (require "tests/stack_test")